requirements.txt
django>=4.2.0,<5.0.0
requests
bs4
djangorestframework
djangorestframework-simplejwt
django-cors-headers
numpy